pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Pluggable AI agent for data fetching and chart generation"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "hanku4u" }
]
license = { text = "MIT" }
keywords = ["pydantic-ai", "agent", "data", "charts", "fastapi"]

dependencies = [
    "pydantic-ai>=0.0.20",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "plotly>=6.0.0",
    "kaleido>=0.2.1",
    "pandas>=2.2.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/hanku4u/data-agent"
Repository = "https://github.com/hanku4u/data-agent"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["data_agent", "data_agent.tools", "data_agent.sources", "data_agent.charts"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
